refactor: consolidate home repositories
