+++ app/src/main/kotlin/app/vimusic/android/ui/viewmodels/HomeDiscoveryViewModel.kt
@@ -2,7 +2,7 @@ package app.vimusic.android.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import app.vimusic.android.repositories.HomeDiscoveryRepository
+import app.vimusic.android.repositories.HomeRepository
 import app.vimusic.android.ui.state.LoadState
 import app.vimusic.android.ui.state.launchLoad
 import app.vimusic.android.utils.requireValue
@@ -14,7 +14,7 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class HomeDiscoveryViewModel(
-    private val repository: HomeDiscoveryRepository
+    private val repository: HomeRepository
 ) : ViewModel() {
     private val mutableUiState = MutableStateFlow<LoadState<YoutubeMusicInnertube.DiscoverPage>>(
         LoadState.Idle
@@ -43,7 +43,7 @@ class HomeDiscoveryViewModel(
     }
 
     companion object {
-        fun factory(repository: HomeDiscoveryRepository) = viewModelFactory {
+        fun factory(repository: HomeRepository) = viewModelFactory {
             HomeDiscoveryViewModel(repository = repository)
         }
     }
