+++ app/src/main/kotlin/app/vimusic/android/ui/viewmodels/HomePlaylistsViewModel.kt
@@ -1,12 +1,12 @@
 package app.vimusic.android.ui.viewmodels
 
 import androidx.lifecycle.ViewModel
-import app.vimusic.android.repositories.HomePlaylistsRepository
+import app.vimusic.android.repositories.HomeRepository
 import app.vimusic.core.data.enums.PlaylistSortBy
 import app.vimusic.core.data.enums.SortOrder
 
 class HomePlaylistsViewModel(
-    private val repository: HomePlaylistsRepository
+    private val repository: HomeRepository
 ) : ViewModel() {
     fun observePlaylistPreviews(sortBy: PlaylistSortBy, sortOrder: SortOrder) =
         repository.observePlaylistPreviews(sortBy = sortBy, sortOrder = sortOrder)
@@ -16,7 +16,7 @@ class HomePlaylistsViewModel(
     fun createPlaylist(name: String) = repository.createPlaylist(name)
 
     companion object {
-        fun factory(repository: HomePlaylistsRepository) = viewModelFactory {
+        fun factory(repository: HomeRepository) = viewModelFactory {
             HomePlaylistsViewModel(repository = repository)
         }
     }
